% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ct_get_links}
\alias{ct_get_links}
\title{ct_get_links}
\usage{
ct_get_links(
  x = "",
  platforms = "",
  count = 100,
  startDate = "",
  endDate = "",
  token = ""
)
}
\arguments{
\item{x}{A number}

\item{platforms}{What platforms}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
ct_get_links
}
