% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ct_get_posts}
\alias{ct_get_posts}
\title{ct_get_posts}
\usage{
ct_get_posts(
  listID = "",
  searchTerm = "",
  language = "",
  types = "",
  minInteractions = 0,
  count = 100,
  startDate = "",
  endDate = "",
  token = ""
)
}
\arguments{
\item{x}{A number}

\item{platforms}{What platforms}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
ct_get_posts
}
\examples{
# add(1, 1)
}
