% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{ct_search_posts}
\alias{ct_search_posts}
\title{ct_search_posts}
\usage{
ct_search_posts(
  x = "",
  and = "",
  not = "",
  inAccountIds = "",
  inListIds = "",
  notInAccountIds = "",
  notInListIds = "",
  notInTitle = "",
  platforms = "",
  types = "",
  minInteractions = 0,
  minSubscriberCount = 0,
  verifiedOnly = "false",
  count = 100,
  startDate = "",
  endDate = "",
  token = ""
)
}
\arguments{
\item{x}{A number}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
ct_search_posts
}
\examples{
# add(1, 1)
}
